finally finished

+++ js/workers/scraperWorker.js
@@ -19,9 +19,9 @@ const JSON_FILE_NAME = path.join(__dirname, '..', '..', 'logs', 'SiteMap_matches
 const ERROR_LOG_NAME = path.join(__dirname, '..', '..', 'logs', 'error_log_');
 const TMP_STORAGE_PATH = path.join(__dirname, '..', '..', 'tmp');
 
-const MAX_CONCURRENT_PROMISES = 5;
+const MAX_CONCURRENT_PROMISES = 4;
 
-const statusTypes = {
+const STATUS_TYPES = {
   0: 'starting',
   1: 'working',
   2: 'done',
@@ -62,45 +62,50 @@ const validateFolder = async (folderName, folderIncludes) => {
   return re.test(folderName);
 };
 
-const handleFile = async (entryName, tags, regex) => {
-  if (path.extname(entryName) === '.pdf') {
-    return scraperService.getRegexMatchesFromPdf(path.join(__dirname, '..', '..', 'tmp', entryName), tags, regex);
+const handleFile = async (fileName, tags, regex) => {
+  if (path.extname(fileName) === '.pdf') {
+    return scraperService.getRegexMatchesFromPdf(path.join(TMP_STORAGE_PATH, fileName), tags, regex); // eslint-disable-line max-len
   }
-  if (path.extname(entryName) === '.docx') {
-    return scraperService.getRegexMatchesFromDocx(path.join(__dirname, '..', '..', 'tmp', entryName), tags, regex);
+  if (path.extname(fileName) === '.docx') {
+    return scraperService.getRegexMatchesFromDocx(path.join(TMP_STORAGE_PATH, fileName), tags, regex); // eslint-disable-line max-len
   }
-  if (path.extname(entryName) === '.txt') {
-    return scraperService.getRegexMatchesFromTxt(path.join(__dirname, '..', '..', 'tmp', entryName), tags, regex);
+  if (path.extname(fileName) === '.txt') {
+    return scraperService.getRegexMatchesFromTxt(path.join(TMP_STORAGE_PATH, fileName), tags, regex); // eslint-disable-line max-len
   }
-  if (path.extname(entryName) === '.kmz') {
-    return scraperService.getDataFromKmz(path.join(__dirname, '..', '..', 'tmp', entryName));
+  if (path.extname(fileName) === '.kmz') {
+    return scraperService.getDataFromKmz(path.join(TMP_STORAGE_PATH, fileName));
   }
   return null;
 };
 
 const handleEntry = async (entry) => {
-  let fileName = '';
-  if (path.dirname(entry.name) !== '.') {
-    if (!(await validateFolder(path.dirname(entry.name), ZIP_DATA.folderIncludes))) return;
-    fileName = path.basename(entry.name);
-  } else {
-    fileName = entry.name;
-  }
-  if (!(await validateFile(fileName, ZIP_DATA.fileIncludes, ZIP_DATA.fileTypes))) return;
-  logger.info(fileName);
-  NUM_FILES_SCRAPED += 1;
-  parentPort.postMessage([NUM_FILES_SCRAPED, statusTypes[1]]);
+  let fileName = null;
+  let folderName = null;
   try {
+    if (path.dirname(entry.name) !== '.') {
+      if (!(await validateFolder(path.dirname(entry.name), ZIP_DATA.folderIncludes))) return;
+      const paths = path.dirname(entry.name).split('/');
+      folderName = paths[paths.length - 1];
+      fileName = `${Date.now()}_${folderName}_${path.basename(entry.name)}`;
+    } else {
+      folderName = (entry.name).replace(/\//g, '');
+      fileName = `${Date.now()}_${(entry.name).replace(/\//g, '')}`;
+    }
+    if (!fileName) return;
+    if (!(await validateFile(fileName, ZIP_DATA.fileIncludes, ZIP_DATA.fileTypes))) return;
+    logger.info(entry.name);
+    NUM_FILES_SCRAPED += 1;
+    parentPort.postMessage([NUM_FILES_SCRAPED, STATUS_TYPES[1]]);
     await ZIP_DATA.zip.extract(entry.name, path.join(TMP_STORAGE_PATH, fileName));
     const matches = await handleFile(fileName, ZIP_DATA.tags, ZIP_DATA.regex);
     if (!matches || matches.length === 0) {
       throw new Error('no matches found');
     }
-    await appendFile(ZIP_DATA.jsonFileName, `"${entry.name}": ${JSON.stringify(matches, null, 4)},`);
+    await appendFile(ZIP_DATA.jsonFileName, `"${folderName}": ${JSON.stringify(matches, null, 4)},`);
   } catch (err) {
-    await appendFile(ZIP_DATA.errFileName, `${ZIP_DATA.filePath},${err.message}\n`);
+    await appendFile(ZIP_DATA.errFileName, `${fileName},${err.message}\n`);
   }
-  await viewerService.removeFile(path.join(TMP_STORAGE_PATH, fileName));
+  if (fileName) await viewerService.removeFile(path.join(TMP_STORAGE_PATH, fileName));
 };
 
 const scrapeZip = async (zipFilePath) => {
@@ -131,7 +136,7 @@ const scrapeZip = async (zipFilePath) => {
   // tell the scraper we're done
   parentPort.postMessage([
     NUM_FILES_SCRAPED,
-    statusTypes[2],
+    STATUS_TYPES[2],
     ZIP_DATA.jsonFileName,
     ZIP_DATA.errFileName,
   ]);
@@ -139,7 +144,7 @@ const scrapeZip = async (zipFilePath) => {
 
 (async () => {
   if (!parentPort) return;
-  parentPort.postMessage([NUM_FILES_SCRAPED, statusTypes[0]]);
+  parentPort.postMessage([NUM_FILES_SCRAPED, STATUS_TYPES[0]]);
   if (path.extname(workerData.filePath) === '.zip') {
     ZIP_DATA.filePath = workerData.filePath;
     ZIP_DATA.fileIncludes = workerData.fileIncludes;
